requirements.txt
# Booklet - Multi-Tenant SaaS ERP Platform
# Python Dependencies

# ============================================
# CORE FRAMEWORK
# ============================================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
flask>=3.0.0

# ============================================
# DATABASE
# ============================================
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
aiosqlite>=0.19.0

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
cryptography>=42.0.0
bcrypt>=4.1.0

# ============================================
# VALIDATION & SERIALIZATION
# ============================================
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# ============================================
# TEMPLATING & FRONTEND
# ============================================
jinja2>=3.1.2
markdown>=3.5.0

# ============================================
# HTTP CLIENT
# ============================================
httpx>=0.26.0
requests>=2.31.0

# ============================================
# AI PROVIDERS
# ============================================
google-generativeai>=0.3.0
z-ai-web-dev-sdk>=0.0.16

# ============================================
# DATA PROCESSING
# ============================================
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

# ============================================
# DOCUMENT GENERATION
# ============================================
weasyprint>=60.2
openpyxl>=3.1.0
reportlab>=4.0.0

# ============================================
# EMAIL
# ============================================
fastapi-mail>=1.4.1

# ============================================
# UTILITIES
# ============================================
python-dateutil>=2.8.2
python-multipart>=0.0.6
aiofiles>=23.2.0

# ============================================
# CACHING & QUEUES
# ============================================
redis>=5.0.0
celery>=5.3.0

# ============================================
# TESTING
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ============================================
# MONITORING
# ============================================
sentry-sdk[fastapi]>=1.40.0
